+++ docs/CHANGES.rst
@@ -4,8 +4,17 @@ Changes
 A method to flatten generated JSON data into timed CSV events in support
 of analytic workflows within the `ContentAI Platform <https://www.contentai.io>`__.
 
-1.1
+1.2
 ---
+
+1.2.2
+~~~~~
+- add parsers for `gcp_videointelligence_text_detection`, `comskip_json`, `ibm_max_audio_classifier`, 
+   `gcp_videointelligence_object_tracking`, `gcp_videointelligence_people_detection`
+- improve testing to iterate over known set of data in testing dir
+- fix generator/parser retrieve for whole name matches, not partials
+- add documentation for new types, explicitly call out `person` tag_type
+- update the `dsai_activity_emotions` parser to return tag type `emotion` (matching that of other AWS, Azure parsers)
 
 1.2.1
 ~~~~~
@@ -16,6 +25,9 @@ of analytic workflows within the `ContentAI Platform <https://www.contentai.io>`
 - add unit-testing to package build
 - add command-line / parser input as complement to contentai-driven ENV variables
 - fix bugs around specification of result path or specific generator
+
+1.1
+---
 
 1.1.8
 ~~~~~